+++ azure-communication-ui/azure-communication-ui/src/main/java/com/azure/android/communication/ui/redux/action/LocalParticipantAction.kt
@@ -62,5 +62,5 @@ internal sealed class LocalParticipantAction : Action {
     ) :
         LocalParticipantAction()
 
-    class DisplayNameIsSet(val displayName: String?) : LocalParticipantAction()
+    class DisplayNameIsSet(val displayName: String) : LocalParticipantAction()
 }