+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Participant list mute icon margin correction
 - Long display name truncation on local participant display view
 - Participant drawer is closed when participant is clicked
+- Sort participants, append suffix for local Participant and handle unnamed participant for the Participant List
 
 ## 1.0.0-beta.1 (2021-12-08)
 This is the initial release of Azure Communication UI Library. For more information, please see the [README][read_me] and [QuickStart][documentation].